requirements.txt
twilio==8.12.0
python-dotenv==1.0.1
fastapi==0.110.0
uvicorn==0.27.1
websockets==12.0
openai==1.12.0
python-multipart==0.0.9
pydantic==2.6.1
PyJWT==2.8.0
slowapi==0.1.9 